: ) alive for bot mode

+++ userge/plugins/bot/alive.py
@@ -2,27 +2,76 @@
 
 import asyncio
 from re import search
-
+from userge.utils import rand_array, get_file_id
 from pyrogram import filters
 from pyrogram.errors import BadRequest, Forbidden
-from pyrogram.types import CallbackQuery
+from pyrogram.types import CallbackQuery, InlineKeyboardButton, InlineKeyboardMarkup,
+from userge.core.ext import RawClient
+from userge import Config, Message, userge, get_version, versions
 
-from userge import Config, Message, userge
-
+CACHED_MEDIA = None
 
 @userge.on_cmd("alive", about={"header": "Just For Fun"}, allow_channels=False)
 async def alive_inline(message: Message):
-    bot = await userge.bot.get_me()
-    try:
-        x = await userge.get_inline_bot_results(bot.username, "alive")
-        y = await userge.send_inline_bot_result(
-            chat_id=message.chat.id, query_id=x.query_id, result_id=x.results[0].id
-        )
-    except (Forbidden, BadRequest) as ex:
-        return await message.err(str(ex), del_in=5)
-    await message.delete()
-    await asyncio.sleep(120)
-    await userge.delete_messages(message.chat.id, y.updates[0].id)
+    global CACHED_MEDIA
+    if message.client.is_bot:
+        if Config.ALIVE_MEDIA:
+            url_ = Config.ALIVE_MEDIA.strip()
+            if url_.lower() == "false":
+                await userge.bot.send_message(
+                    message.chat.id,
+                    Bot_Alive.alive_info(),
+                    reply_markup=Bot_Alive.alive_buttons(),
+                    disable_web_page_preview=True
+                )
+            else:
+                type_, media_ = await Bot_Alive.check_media_link(Config.ALIVE_MEDIA)
+                if type_ == "url_gif":
+                    await userge.bot.send_animation(
+                        message.chat.id,
+                        animation=url_,
+                        caption=Bot_Alive.alive_info(),
+                        reply_markup=Bot_Alive.alive_buttons(),   
+                    )
+                elif type_ == "url_image"
+                    await userge.bot.send_photo(
+                        message.chat.id,
+                        photo=url_,
+                        caption=Bot_Alive.alive_info(),
+                        reply_markup=Bot_Alive.alive_buttons(),   
+                    )
+                elif type_ == "tg_media" and isinstance(media_, list):
+                    if not CACHED_MEDIA:
+                        try:
+                            CACHED_MEDIA = get_file_id(await userge.bot.get_messages(media_[0], media_[1]))
+                        except Exception as er:
+                            await message.err(er, del_in=7)
+                            return
+                    await userge.bot.send_cached_media(
+                        message.chat.id,
+                        file_id=CACHED_MEDIA,
+                        caption=Bot_Alive.alive_info(),
+                        reply_markup=Bot_Alive.alive_buttons(), 
+                    )
+        else:
+            await userge.bot.send_photo(
+                message.chat.id,
+                photo=Bot_Alive.alive_default_imgs(),
+                caption=Bot_Alive.alive_info(),
+                reply_markup=Bot_Alive.alive_buttons(),   
+            )
+    else:
+        bot = await userge.bot.get_me()
+        try:
+            x = await userge.get_inline_bot_results(bot.username, "alive")
+            y = await userge.send_inline_bot_result(
+                chat_id=message.chat.id, query_id=x.query_id, result_id=x.results[0].id
+            )
+        except (Forbidden, BadRequest) as ex:
+            return await message.err(str(ex), del_in=5)
+        await message.delete()
+        await asyncio.sleep(120)
+        await userge.delete_messages(message.chat.id, y.updates[0].id)
 
 
 if userge.has_bot:
@@ -46,24 +95,69 @@ def _parse_arg(arg: bool) -> str:
     return "  ✅  Enabled" if arg else "  ❌  Disabled"
 
 
-async def check_media_link(media_link: str):
-    alive_regex_ = r"http[s]?://(i\.imgur\.com|telegra\.ph/file|t\.me)/(\w+)(?:\.|/)(gif|jpg|png|jpeg|[0-9]+)(?:/([0-9]+))?"
-    match = search(alive_regex_, media_link)
-    if not match:
-        return None, None
-    if match.group(1) == "i.imgur.com":
-        link = match.group(0)
-        link_type = "url_gif" if match.group(3) == "gif" else "url_image"
-    elif match.group(1) == "telegra.ph/file":
-        link = match.group(0)
-        link_type = "url_image"
-    else:
-        link_type = "tg_media"
-        if match.group(2) == "c":
-            chat_id = int("-100" + str(match.group(3)))
-            message_id = match.group(4)
+class Bot_Alive:
+    @staticmethod
+    async def check_media_link(media_link: str):
+        alive_regex_ = r"http[s]?://(i\.imgur\.com|telegra\.ph/file|t\.me)/(\w+)(?:\.|/)(gif|jpg|png|jpeg|[0-9]+)(?:/([0-9]+))?"
+        match = search(alive_regex_, media_link)
+        if not match:
+            return None, None
+        if match.group(1) == "i.imgur.com":
+            link = match.group(0)
+            link_type = "url_gif" if match.group(3) == "gif" else "url_image"
+        elif match.group(1) == "telegra.ph/file":
+            link = match.group(0)
+            link_type = "url_image"
         else:
-            chat_id = match.group(2)
-            message_id = match.group(3)
-        link = [chat_id, int(message_id)]
-    return link_type, link
+            link_type = "tg_media"
+            if match.group(2) == "c":
+                chat_id = int("-100" + str(match.group(3)))
+                message_id = match.group(4)
+            else:
+                chat_id = match.group(2)
+                message_id = match.group(3)
+            link = [chat_id, int(message_id)]
+        return link_type, link
+    
+    @staticmethod
+    def alive_info():
+        alive_info = f"""
+<b>[USERGE-X](https://telegram.dog/x_xtests)  is Up and Running.
+
+    Python</b> :                 🐍  <code>v{versions.__python_version__}</code>
+    <b>Pyrogram</b> :         🔥  <code>v{versions.__pyro_version__}</code>
+    𝑿 :        🧬  <code>v{get_version()}</code>
+
+⚙ <b>Mode  ➥    [ {Bot_Alive._get_mode()} ]</b>
+"""
+        return alive_info
+    
+    @staticmethod
+    def _get_mode() -> str:
+        if RawClient.DUAL_MODE:
+            return "↕️  DUAL"
+        if Config.BOT_TOKEN:
+            return "🤖  BOT"
+        return "👤  USER"
+
+    @staticmethod
+    def alive_buttons():
+        buttons = [
+            [
+                InlineKeyboardButton("SETTINGS", callback_data="settings_btn"),
+                InlineKeyboardButton(text="REPO", url=Config.UPSTREAM_REPO),
+            ]
+        ]
+        return InlineKeyboardMarkup(buttons)
+
+    @staticmethod
+    def alive_default_imgs():
+        alive_imgs = [
+            "https://telegra.ph/file/11123ef7dff2f1e19e79d.jpg",
+            "https://i.imgur.com/uzKdTXG.jpg",
+            "https://telegra.ph/file/6ecab390e4974c74c3764.png",
+            "https://telegra.ph/file/995c75983a6c0e4499b55.png",
+            "https://telegra.ph/file/86cc25c78ad667ca5e691.png",
+        ]
+        return rand_array(alive_imgs)
+
